+++ src/lib.rs
@@ -1,6 +1,7 @@
 #![recursion_limit = "1024"]
 
 pub mod bindings;
+pub mod embedded_stdlib;
 pub mod module_loader;
 pub mod resolver;
 pub mod worker;